+++ src/fable-library-ts/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+* [JS/TS] Fixed quotation for union string cases (by @MangelMaxime)
+
 ## 2.0.0-rc.4 - 2026-03-19
 
 ### Fixed